@@ -1,35 +1,30 @@
 [package]
-name    = "stored"
-version = "0.1.0"
-authors = ["Thomas de Zeeuw <thomasdezeeuw@gmail.com>"]
-edition = "2018"
-publish = false
+name        = "stored"
+description = "Store*d* (pronounced store-daemon, or just stored) is a distributed immutable blob store. Stored is not a key-value store, as the key isn't the decided by the user but by the SHA-512 checksum of the stored blob."
+version     = "0.2.0-dev"
+publish     = false # In development.
+authors     = ["Thomas de Zeeuw <thomasdezeeuw@gmail.com>"]
+license     = "MIT"
+repository  = "https://github.com/Thomasdezeeuw/stored"
+readme      = "README.md"
+edition     = "2021"
 
 [dependencies]
-chrono            = { version = "0.4.19", default-features = false, features = ["clock"] }
-crossbeam-channel = { version = "0.5.0", default-features = false, features = ["std"] }
-futures-io        = { version = "0.3.6", default-features = false, features = ["std"] }
-futures-util      = { version = "0.3.6", default-features = false, features = ["io", "std", "write-all-vectored"] }
-fxhash            = { version = "0.2.1", default-features = false }
-getrandom         = { version = "0.2.0", default-features = false }
-# TODO: replace with published crate.
-heph              = { git = "https://github.com/Thomasdezeeuw/heph", rev = "00640b2a5d1759b572a2c9e9e13b06df4f3e40c5" }
-libc              = { version = "0.2.74", default-features = false }
-log               = { version = "0.4.11", default-features = false, features = ["release_max_level_info"] }
-num_cpus          = { version = "1.13.0", default-features = false }
-ring              = { version = "0.16.19", default-features = false, features = ["std"] }
-# Need the `std` feature as it implements `Error` for `httparse::Error`.
-httparse          = { version = "1.3.4", default-features = false, features = ["std"] }
-human-size        = { version = "0.4.1", default-features = false, features = ["enable-serde"] }
-parking_lot       = { version = "0.11.0", default-features = false, features = ["nightly"] }
-serde             = { version = "1.0.116", default-features = false, features = ["derive"] }
-serde_json        = { version = "1.0.58", default-features = false, features = ["std"] }
-toml              = { version = "0.5.6", default-features = false }
+# TODO: replace with published crates.
+left-right        = { git = "https://github.com/Thomasdezeeuw/left-right", rev = "4408ba0bb23ea0090b420ebfc0561780a44c533f", default-features = false }
+heph              = { git = "https://github.com/Thomasdezeeuw/heph", rev = "9aa893cc1089124b515b27a910303e60d1ba8960", default-features = false }
+heph-rt           = { git = "https://github.com/Thomasdezeeuw/heph", rev = "9aa893cc1089124b515b27a910303e60d1ba8960", default-features = false }
+heph-http         = { git = "https://github.com/Thomasdezeeuw/heph", rev = "9aa893cc1089124b515b27a910303e60d1ba8960", default-features = false }
+itoa              = { version = "1.0.4", default-features = false }
+libc              = { version = "0.2.151", default-features = false }
+log               = { version = "0.4.17", default-features = false, features = ["kv_unstable_std", "release_max_level_info"] }
+ring              = { version = "0.17", default-features = false }
+std-logger        = { version = "0.5", default-features = false, features = ["timestamp", "log-panic", "nightly"] }
 
 [dev-dependencies]
 http        = { version = "0.2.1", default-features = false }
 lazy_static = { version = "1.4.0", default-features = false }
-mio         = { version = "0.7.0", default-features = false, features = ["os-poll"] }
+mio         = { version = "0.8.0", default-features = false, features = ["os-poll"] }
 mio-pipe    = { version = "0.1.1", default-features = false }
 serde_test  = { version = "1.0.116", default-features = false }
 
